example/backend/admin/requirements.txt
Django==3.2.4
djangorestframework==3.12.4
mysqlclient==2.0.3
django-mysql==3.11.1
django-cors-headers==3.7.0
pika==1.2.0